Fixed the bug that was mangling higher resolution polylines.
New bug to fix: script now mistakenly reads Skeinforge comments at the wrong time and inserts blank 'layer' commands
in the wrong places as a result. It's finding different Z pos values in the <boundaryPoint> comments and therefore inserting
layer commands. It should exclude this check if the current line is a Skeinforge comment.